refactor: 강의 정보 전역 스토어에서 정보 가져오는 형태로 변경
- 추후 API 요청을 통해 과목들을 저장할 예정임
- 따라서 구조 변경
